+++ tests/integration/Mocks/PayoutResponse.php
@@ -23,7 +23,7 @@ class PayoutResponse
         return new Response(200, [], '{"id":"cbc98b01-6d4e-4137-a6fc-1659a26e5c55","merchant_account_id":"822f8dfe-0874-481d-b966-5b14f767792f","amount_in_minor":1,"currency":"GBP","beneficiary":{"type":"external_account","account_holder_name":"Test","account_identifier":{"type":"sort_code_account_number","sort_code":"121212","account_number":"12345678"},"reference":"Test reference"},"status":"executed","created_at":"2022-04-01T19:47:12.642232Z","authorized_at":"2022-04-01T19:47:12.642232Z","executed_at":"2022-04-01T19:47:13.642232Z"}');
     }
 
-    public static function executed_with_metadata(): Response
+    public static function executedWithMetadata(): Response
     {
         return new Response(200, [], '{"id":"cbc98b01-6d4e-4137-a6fc-1659a26e5c55","merchant_account_id":"822f8dfe-0874-481d-b966-5b14f767792f","amount_in_minor":1,"currency":"GBP","beneficiary":{"type":"external_account","account_holder_name":"Test","account_identifier":{"type":"sort_code_account_number","sort_code":"121212","account_number":"12345678"},"reference":"Test reference"},"status":"executed","created_at":"2022-04-01T19:47:12.642232Z","authorized_at":"2022-04-01T19:47:12.642232Z","executed_at":"2022-04-01T19:47:13.642232Z", "metadata": {"foo": "bar"}}');
     }