+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Support for metadata field on payout creation, retrieval and webhooks (#68)
+- Support for metadata field on refund creation, retrieval and webhooks (#68)
+
+
 ## [2.6.0] - 2024-10-10
 
 ### Added